ocr/requirements.txt
pytesseract
pdf2image
psycopg2-binary
meilisearch
requests
Pillow
python-dotenv
